backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
supabase>=2.3.0
google-generativeai>=0.3.0
exa-py>=1.0.0
tavily-python
